Project1/requirements.txt
Flask
psycopg2-binary
python-dotenv
watchdog
